Move envelope classes under the `envelope` package.
