+++ server/src/main/java/org/spine3/server/command/CommandBus.java
@@ -26,7 +26,6 @@ import io.grpc.stub.StreamObserver;
 import org.spine3.Internal;
 import org.spine3.base.Command;
 import org.spine3.base.CommandClass;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.CommandId;
 import org.spine3.base.Error;
 import org.spine3.base.Errors;
@@ -34,6 +33,7 @@ import org.spine3.base.FailureThrowable;
 import org.spine3.base.Identifiers;
 import org.spine3.base.Response;
 import org.spine3.base.Responses;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.BoundedContext;
 import org.spine3.server.Statuses;
 import org.spine3.server.bus.Bus;
