+++ server/src/main/java/org/spine3/server/command/CommandHandler.java
@@ -26,8 +26,8 @@ import com.google.protobuf.Message;
 import com.google.protobuf.StringValue;
 import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.Event;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.protobuf.AnyPacker;
 import org.spine3.server.event.EventBus;
 import org.spine3.server.reflect.CommandHandlerMethod;
