+++ server/src/main/java/org/spine3/server/stand/StandFunnel.java
@@ -20,6 +20,7 @@
 package org.spine3.server.stand;
 
 import org.spine3.Internal;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.entity.VersionableEntity;
 
 import static com.google.common.base.Preconditions.checkNotNull;
@@ -40,7 +41,7 @@ import static com.google.common.base.Preconditions.checkState;
  * there can be the only instance of {@code StandFunnel}.
  *
  * @author Alex Tymchenko
- * @see org.spine3.server.aggregate.AggregateRepository#dispatch(org.spine3.base.CommandEnvelope)
+ * @see org.spine3.server.aggregate.AggregateRepository#dispatch(CommandEnvelope)
  *      AggregateRepository.dispatch(Command)
  * @see org.spine3.server.projection.ProjectionRepository#dispatch(org.spine3.base.Event)
  *      ProjectionRepository.dispatch(Event)
