+++ server/src/test/java/org/spine3/server/command/CommandBusShould.java
@@ -26,10 +26,11 @@ import org.junit.Test;
 import org.spine3.base.Command;
 import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.CommandValidationError;
 import org.spine3.base.Error;
 import org.spine3.base.Response;
+import org.spine3.envelope.CommandEnvelope;
+import org.spine3.envelope.MessageEnvelope;
 import org.spine3.server.command.error.CommandException;
 import org.spine3.server.command.error.InvalidCommandException;
 import org.spine3.server.command.error.UnsupportedCommandException;
@@ -357,7 +358,7 @@ public class CommandBusShould extends AbstractCommandBusTestSuite {
 
     /**
      * The dispatcher that remembers that
-     * {@link CommandDispatcher#dispatch(org.spine3.base.MessageEnvelope) dispatch()} was called.
+     * {@link CommandDispatcher#dispatch(MessageEnvelope) dispatch()} was called.
      */
     private static class AddTaskDispatcher implements CommandDispatcher {
