+++ server/src/test/java/org/spine3/server/command/CommandBusShouldHandleCommandStatus.java
@@ -30,12 +30,12 @@ import org.junit.Test;
 import org.spine3.base.Command;
 import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.CommandId;
 import org.spine3.base.Errors;
 import org.spine3.base.FailureThrowable;
 import org.spine3.base.Response;
 import org.spine3.client.CommandFactory;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.protobuf.Durations2;
 import org.spine3.server.event.EventBus;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
