+++ server/src/test/java/org/spine3/server/command/CommandHandlerShould.java
@@ -29,11 +29,11 @@ import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.spine3.base.Command;
 import org.spine3.base.CommandContext;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.CommandId;
 import org.spine3.base.Commands;
 import org.spine3.base.Event;
 import org.spine3.base.Events;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.event.EventBus;
 import org.spine3.test.command.AddTask;
 import org.spine3.test.command.CreateProject;
