+++ server/src/test/java/org/spine3/server/command/CommandSchedulingShould.java
@@ -25,7 +25,7 @@ import com.google.protobuf.Timestamp;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.spine3.base.Command;
-import org.spine3.base.CommandEnvelope;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.protobuf.Durations2;
 
 import java.util.List;
