+++ server/src/test/java/org/spine3/server/procman/ProcessManagerRepositoryShould.java
@@ -32,11 +32,11 @@ import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.spine3.base.CommandClass;
 import org.spine3.base.CommandContext;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.Event;
 import org.spine3.base.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.base.Events;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.server.BoundedContext;
 import org.spine3.server.command.Assign;
 import org.spine3.server.command.CommandDispatcher;
