+++ server/src/test/java/org/spine3/server/stand/StandFunnelShould.java
@@ -26,9 +26,9 @@ import io.netty.util.internal.ConcurrentSet;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.ArgumentMatchers;
-import org.spine3.base.CommandEnvelope;
 import org.spine3.base.Identifiers;
 import org.spine3.base.Version;
+import org.spine3.envelope.CommandEnvelope;
 import org.spine3.protobuf.AnyPacker;
 import org.spine3.protobuf.Timestamps2;
 import org.spine3.server.BoundedContext;
